+++ examples/document_code_base/workflows/document_file.yml
@@ -52,6 +52,9 @@
             Do not add a logger into the `__init__.py`. If one exists within
             there and it's not being used, remove it.
 
+            Do NOT add additional logging into the file. Just make sure any
+            existing logging uses the log and this approach.
+
 
             METHODS OR FUNCTIONS
             --------------------
@@ -117,7 +120,7 @@
             The filename is: `{{ params.filename }}`
 
             [--------------- FILE CONTENT: START ---------- ]
-            {{ params.content }}
+            {{{ params.content }}}
             [--------------- FILE CONTENT: END ------------ ]
 
       response_format:
