+++ src/reasonchip/core/engine/README.md
@@ -3,12 +3,3 @@
 The Engine is the core part of the system responsible for loading
 and running of pipelines.
 
-## Filesystem layout
-
-| Location | Description |
-| ------------------------- | ----------------------------------------------- |
-| [context](./context/) | Context classes |
-| [parsers](./parsers/) | Parser locations |
-| [processor](./processor/) | The processor |
-
-
