--- src/reasonchip/core/engine/context/__init__.py
@@ -1,13 +1,0 @@
-# SPDX-License-Identifier: GPL-3.0-or-later
-# Copyright (C) 2025 South Patron LLC
-# This file is part of ReasonChip and licensed under the GPLv3+.
-# See <https://www.gnu.org/licenses/> for details.
-
-from .flow_control import FlowControl
-from .variables import Variables, VariableMapType
-
-
-__all__ = [
-    "FlowControl",
-    "Variables",
-]
