--- src/reasonchip/core/engine/parsers/__init__.py
@@ -1,4 +1,0 @@
-# SPDX-License-Identifier: GPL-3.0-or-later
-# Copyright (C) 2025 South Patron LLC
-# This file is part of ReasonChip and licensed under the GPLv3+.
-# See <https://www.gnu.org/licenses/> for details.
